backend/requirements.txt
Flask==2.3.3
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
matplotlib==3.7.1
flask-cors==4.0.0
